+++ pkg/providers/cloudcertificates/testdata/TestDataCertificates/with_expiring_in_days_positive_value.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudcertificates_certificates" "test" {
+  expiring_in_days = 980
+}