+++ pkg/providers/cloudcertificates/testdata/TestDataCertificates/with_key_type_and_sort.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudcertificates_certificates" "test" {
+  key_type = "RSA"
+  sort     = "-createdDate"
+}